Overridden nav bar back buttons to only finish the activity rather start parent

+++ app/src/main/java/com/example/james/todolist/activities/TaskViewerActivity.java
@@ -2,6 +2,7 @@ package com.example.james.todolist.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -47,6 +48,10 @@ public class TaskViewerActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item){
+        if(item.getItemId() == android.R.id.home){
+            finish();
+            return true;
+        }
         if(item.getItemId() == R.id.bin_icon_task_viewer_activity){
             currentTask.delete();
             currentTask = null;
